fix: CRITICAL ERROR checkout only working for logged in users. sendMail logic moved to service function (mails + notifications)
